+++ src/app/builders/components/Table/Cell/BackingShareCell/index.ts
@@ -1,1 +1,1 @@
-export * from './BackingShare'
+export * from './BackingShareCell'
